requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
beautifulsoup4==4.6.0
cffi==1.11.2
Chameleon==3.2
colander==1.4
ColanderAlchemy==0.3.3
coverage==4.4.2
cryptography==2.1.3
deform==2.0.4
hupper==1.0
idna==2.6
iso8601==0.1.12
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.14.0
paginate==0.5.6
paginate-sqlalchemy==0.2.0
PasteDeploy==1.5.2
peppercorn==0.5
pkg-resources==0.0.0
plaster==1.0
plaster-pastedeploy==0.4.2
pluggy==0.6.0
pprint==0.1
ps-alchemy==0.0.7
ps-crud==0.0.4
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.18
pydash==4.3.0
Pygments==2.2.0
PyJWT==1.5.3
pyramid==1.9.1
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-jwtauth==0.1.3
pyramid-mako==1.0.2
pyramid-retry==0.5
pyramid-sacrud==0.3.3
pyramid-sqlalchemy==1.6
pyramid-tm==2.2
pytest==3.2.5
pytest-cov==2.5.1
python-dateutil==2.7.3
repoze.lru==0.7
sacrud==0.3.7.post0
sacrud-deform==0.1.6
saexttype==0.0.2
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.1.15
transaction==2.1.2
translationstring==1.3
venusian==1.1.0
waitress==1.1.0
WebOb==1.7.4
WebTest==2.0.29
wsgicors==0.7.0
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==0.7.7
